refactor Example
