+++ packages/grammar-sdk/index.ts
@@ -1,6 +1,3 @@
 export * from "./src/grammar";
-export type { GrammarPoint } from "./src/types/GrammarPoint";
 export type { Exercise } from "./src/types/Exercise";
-import type { Example as E } from "./src/types/Example";
-export type Example = E;
-export * as Example from "./src/impl/Example";
+export type { GrammarPoint } from "./src/types/GrammarPoint";
