--- packages/grammar-sdk/src/types/Example.ts
@@ -1,10 +1,0 @@
-export type Example = [string, string, string];
-export const Example = (str: string): Example => {
-  const regex = /(.*)%(.*?)%(.*)/;
-  const matches = str.match(regex);
-
-  if (matches) {
-    return [matches[1] ?? "", matches[2] ?? "", matches[3] ?? ""];
-  }
-  return ["", "", ""];
-};
