+++ packages/grammar-sdk/src/types/GrammarPoint.ts
@@ -1,4 +1,4 @@
-import type { Example } from "./Example";
+import type { Example } from "../example";
 import type { Exercise } from "./Exercise";
 
 export interface GrammarPoint {
